auth_service/requirements.txt
# Core Django and REST Framework
Django>=5.1.7,<5.2
djangorestframework>=3.15.2,<3.16
dj-rest-auth>=7.0.1,<7.2
# django-allauth>=65.4.1,<65.6
djangorestframework-simplejwt>=5.5.0,<5.6
django-cors-headers>=4.7.0,<4.8

# Database
psycopg2-binary>=2.9.10,<2.10

# Image Handling (for profile_picture)
Pillow>=11.1.0,<11.2

# Social Authentication
social-auth-app-django>=5.4.2,<5.5
# Django Dependencies
asgiref>=3.8.1,<3.9
sqlparse>=0.5.3,<0.6
tzdata>=2025.1,<2026.0
 # Optional, only if on Windows
requests

